requirements.txt
asgiref==3.4.1
Django==3.2.7
django-rest-framework==0.1.0
djangorestframework==3.12.4
docxcompose==1.3.3
docxtpl==0.12.0
Jinja2==3.0.1
lxml==4.6.3
MarkupSafe==2.0.1
python-docx==0.8.11
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
